biber: update to 2.16, rename installation script
Users will need to run termux-intall-biber manually to install
biber. This hopefully makes the package a bit nicer to deal with,
running into an error in the postinst script without the option to
stop the installation is annoying.

Also fix issue with cpanm not being found. My test environment have had
~/perl5/bin added to path from .bashrc, and so cpanm was always
found. Trying on a fresh installation revealed the issue.

Have also verified that it works in the fish shell.